# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Online Banking Documentation",
        "description": "Online Banking OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1/api/v1",
            "description": "Online Banking API Server"
        }
    ],
    "paths": {
        "/accounts/statement": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Get list of account transactions",
                "description": "Returns list of account's transactions",
                "operationId": "getAccountStatement",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/transfers/batch": {
            "post": {
                "tags": [
                    "Transfers"
                ],
                "summary": "Store new batch transfer",
                "description": "Returns transfers data from a batch",
                "operationId": "storeBatchTransfer",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreBatchTransferRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
        "/subscriptions": {
            "get": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Get list of subscriptions",
                "description": "Returns subscriptions list",
                "operationId": "getSubscriptions",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "path",
                        "description": "Amount of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "path",
                        "description": "Current page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "post": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Store new subscription",
                "description": "Returns subscription data",
                "operationId": "storeSubscription",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreSubscriptionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
        "/transfers": {
            "get": {
                "tags": [
                    "Transfers"
                ],
                "summary": "Get list of transfers",
                "description": "Returns transfers data",
                "operationId": "getTransfers",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "path",
                        "description": "Amount of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "path",
                        "description": "Current page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "post": {
                "tags": [
                    "Transfers"
                ],
                "summary": "Store new transfer",
                "description": "Returns transfer data",
                "operationId": "storeTransfer",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreTransferRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
        "/transfers/{id}": {
            "get": {
                "tags": [
                    "Transfers"
                ],
                "summary": "Get transfer data",
                "description": "Returns transfer data",
                "operationId": "getTransfer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Transfer id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TransferResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "StoreBatchTransferRequest": {
                "title": "StoreBatchTransferRequest",
                "description": "Store BatchTransfer request body data",
                "required": [
                    "batch_id",
                    "items"
                ],
                "properties": {
                    "batch_id": {
                        "title": "batch_id",
                        "description": "Client-generated id of transfers batch",
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000 "
                    },
                    "items": {
                        "title": "items",
                        "description": "Transfers to create",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/StoreTransferRequest"
                        },
                        "example": "[{'reference_id': '550e8400-e29b-41d4-a716-446655440000 ','amount': 100,'recipient_user_id': 2,'description': 'Payment for items'}]"
                    }
                },
                "type": "object"
            },
            "StoreSubscriptionRequest": {
                "title": "StoreSubscriptionRequest",
                "description": "Store Subscription request body data",
                "required": [
                    "name",
                    "price",
                    "frequency",
                    "start_date"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Subscription name/description",
                        "type": "string",
                        "example": "Monthly Oreylis"
                    },
                    "price": {
                        "title": "price",
                        "description": "Cost of subscription",
                        "type": "integer",
                        "example": "100"
                    },
                    "frequency": {
                        "title": "frequency",
                        "description": "Frequency of subscription (monthly/yearly)",
                        "type": "string",
                        "example": "monthly"
                    },
                    "start_date": {
                        "title": "start_date",
                        "description": "Date from which subscription starts",
                        "type": "string",
                        "format": "date",
                        "example": "2024-07-01"
                    }
                },
                "type": "object"
            },
            "StoreTransferRequest": {
                "title": "StoreTransferRequest",
                "description": "Store Transfer request body data",
                "required": [
                    "reference_id",
                    "amount",
                    "recipient_user_id"
                ],
                "properties": {
                    "reference_id": {
                        "title": "reference_id",
                        "description": "Client-generated id of transfer",
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000 "
                    },
                    "amount": {
                        "title": "amount",
                        "description": "Amount of money to transfer",
                        "type": "integer",
                        "example": "100"
                    },
                    "recipient_user_id": {
                        "title": "recipient_user_id",
                        "description": "ID of the user who will receive money",
                        "type": "integer",
                        "example": "2"
                    },
                    "description": {
                        "title": "description",
                        "description": "Short description of transaction puprose",
                        "type": "string",
                        "example": "Payment for items"
                    }
                },
                "type": "object"
            },
            "TransferResource": {
                "title": "TransferResource",
                "description": "Transfer resource",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Transfer ID",
                        "type": "integer",
                        "example": "1"
                    },
                    "batch_id": {
                        "title": "batch_id",
                        "description": "Client-generated id of transfers batch",
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000 "
                    },
                    "reference_id": {
                        "title": "reference_id",
                        "description": "Client-generated id of transfer",
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000 "
                    },
                    "amount": {
                        "title": "amount",
                        "description": "Amount of money to transfer",
                        "type": "integer",
                        "example": "100"
                    },
                    "recipient_user_id": {
                        "title": "recipient_user_id",
                        "description": "ID of the user who will receive money",
                        "type": "integer",
                        "example": "2"
                    },
                    "description": {
                        "title": "description",
                        "description": "Short description of transaction puprose",
                        "type": "string",
                        "example": "Payment for items"
                    },
                    "status": {
                        "title": "status",
                        "description": "Status of the transfer",
                        "type": "string",
                        "example": "declined"
                    },
                    "status_message": {
                        "title": "status_message",
                        "description": "Status message of the transfer",
                        "type": "string",
                        "example": "Not enough money on the sender's balance."
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "SDate and time of the transfer creation",
                        "type": "string",
                        "format": "date",
                        "example": "2024-07-01"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Accounts",
            "description": "Accounts"
        },
        {
            "name": "Transfers",
            "description": "Transfers"
        },
        {
            "name": "Subscriptions",
            "description": "Subscriptions"
        }
    ]
}